# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Albums API service Swagger Documentation",
        "version": "1.0.0",
        "title": "Albums API"
    },
    "schemes": [
        "https"
    ],
    "host": "albums-collection-service.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/albums": {
            "get": {
                "summary": "Get all the albums",
                "description": "Get all the albums",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AlbumResponseDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create an album",
                "description": "Create an album",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "album object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/AlbumPostRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/CreateAlbumResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/albums/{id}": {
            "get": {
                "summary": "Get an album",
                "description": "Get an album",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Get album by album_id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/AlbumResponseDto"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid status value",
                        "message": "Not found"
                    }
                }
            },
            "put": {
                "summary": "Update an album",
                "description": "Update an album",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "album_id that needs to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "album object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/AlbumPutRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/UpdateAlbumResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update an album",
                "description": "Update an album",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "album_id that needs to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "album object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/AlbumPatchRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/UpdateAlbumResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete an album",
                "description": "Delete an album",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "album_id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/DeleteAlbumResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "AlbumPostRequestDto": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "required": true
                },
                "artist": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "songs": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "AlbumPatchRequestDto": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "songs": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "AlbumPutRequestDto": {
            "type": "object",
            "required": [
                "title",
                "artist",
                "genre",
                "label",
                "songs",
                "year"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "required": true
                },
                "artist": {
                    "type": "string",
                    "required": true
                },
                "genre": {
                    "type": "string",
                    "required": true
                },
                "label": {
                    "type": "string",
                    "required": true
                },
                "songs": {
                    "type": "integer",
                    "required": true
                },
                "year": {
                    "type": "integer",
                    "required": true
                }
            }
        },
        "AlbumResponseDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "album_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "songs": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "CreateAlbumResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "album_id": {
                    "type": "string"
                }
            }
        },
        "UpdateAlbumResponse": {
            "type": "object",
            "properties": {
                "acknowledged": {
                    "type": "boolean"
                },
                "modifiedCount": {
                    "type": "integer"
                },
                "upsertedId": {
                    "type": "object"
                },
                "modifiupsertedCountedCount": {
                    "type": "integer"
                },
                "matchedCount": {
                    "type": "integer"
                }
            }
        },
        "DeleteAlbumResponse": {
            "type": "object",
            "properties": {
                "acknowledged": {
                    "type": "boolean"
                },
                "deletedCount": {
                    "type": "integer"
                }
            }
        }
    }
}